+++ Modules/ADUtils/NavigationControllerObserver.swift
@@ -7,6 +7,7 @@
 //
 
 import Foundation
+import UIKit
 
 /**
  Protocol NavigationControllerObserverDelegate provides methods to observe `NavigationControllerObserver` actions
@@ -55,13 +56,16 @@ public class NavigationControllerObserver : NSObject, UINavigationControllerDele
         didSet {
             // Make the navigationController reevaluate respondsToSelector
             // for UINavigationControllerDelegate methods
-            navigationController.delegate = nil
-            navigationController.delegate = self
+            navigationController?.delegate = nil
+            navigationController?.delegate = self
         }
     }
 
-    private var viewControllersToDelegates: [UIViewController: NavigationControllerObserverDelegateContainer] = [:]
-    private let navigationController: UINavigationController
+    // Do not retain view controllers, but retain delegate containers as they are containers to weak objects, or they
+    // will be deallocated right the way
+    private let viewControllersToDelegates = NSMapTable<UIViewController, NavigationControllerObserverDelegateContainer>
+        .weakToStrongObjects()
+    private weak var navigationController: UINavigationController?
 
     /**
      - parameter navigationController: the `UINavigationController` we want to observe
@@ -98,7 +102,7 @@ public class NavigationControllerObserver : NSObject, UINavigationControllerDele
     public func observePopTransition(of viewController: UIViewController,
                                      delegate: NavigationControllerObserverDelegate) {
         let wrappedDelegate = NavigationControllerObserverDelegateContainer(value: delegate)
-        viewControllersToDelegates[viewController] = wrappedDelegate
+        viewControllersToDelegates.setObject(wrappedDelegate, forKey: viewController)
     }
 
     /**
@@ -106,7 +110,7 @@ public class NavigationControllerObserver : NSObject, UINavigationControllerDele
      - parameter viewController: The viewController being observed
      */
     public func removeDelegate(observing viewController: UIViewController) {
-        viewControllersToDelegates.removeValue(forKey: viewController)
+        viewControllersToDelegates.removeObject(forKey: viewController)
     }
 
     /**
@@ -115,11 +119,10 @@ public class NavigationControllerObserver : NSObject, UINavigationControllerDele
      */
     public func remove(_ delegate: NavigationControllerObserverDelegate) {
         let viewControllersToRemove = viewControllersToDelegates
-            .filter {
-                let (_, wrappedDelegate) = $0
-                return wrappedDelegate.value === delegate
-            }
-            .keys
+            .keyEnumerator()
+            .allObjects
+            .compactMap { $0 as? UIViewController }
+            .filter { viewControllersToDelegates.object(forKey: $0)?.value === delegate }
         viewControllersToRemove.forEach { removeDelegate(observing: $0) }
     }
 
@@ -127,7 +130,7 @@ public class NavigationControllerObserver : NSObject, UINavigationControllerDele
      Stop observing all the viewController and remove all delegates
      */
     public func removeAllDelegates() {
-        viewControllersToDelegates.removeAll()
+        viewControllersToDelegates.removeAllObjects()
     }
 
     //MARK: - UINavigationControllerDelegate
@@ -148,7 +151,7 @@ public class NavigationControllerObserver : NSObject, UINavigationControllerDele
         }
 
         // Clean up state before calling delegate method
-        let delegateToCall = viewControllersToDelegates[fromViewController]?.value
+        let delegateToCall = viewControllersToDelegates.object(forKey: fromViewController)?.value
         removeDelegate(observing: fromViewController)
         cleanOutdatedViewControllers()
         delegateToCall?.navigationControllerObserver(self, didObservePopTransitionFor: fromViewController)
@@ -164,9 +167,11 @@ public class NavigationControllerObserver : NSObject, UINavigationControllerDele
     }
 
     private func cleanOutdatedViewControllers() {
-        let viewControllersToRemove = viewControllersToDelegates.keys.filter {
-            $0.parent != self.navigationController
-        }
+        let viewControllersToRemove = viewControllersToDelegates
+            .keyEnumerator()
+            .allObjects
+            .compactMap { $0 as? UIViewController }
+            .filter { $0.parent != self.navigationController }
         viewControllersToRemove.forEach { removeDelegate(observing: $0) }
     }
 }